replace pages in directory pages
