+++ src/components/Forms/FormPasswordReset.jsx
@@ -1,22 +1,15 @@
 import React, { useEffect } from 'react';
-import { useNavigate, useParams } from 'react-router-dom';
+import { useParams } from 'react-router-dom';
 import { useDispatch, useSelector } from 'react-redux';
 import { Formik, Form, useFormikContext } from 'formik';
 import InputPassword from '../Inputs/InputPassword';
 import { fetchConfirmPasswordReset } from '../../services/thunks/resetPasswordThunk';
-import { openModal } from '../../services/slices/modalSlice';
 import { clearConfirmPasswordError } from '../../services/slices/resetPasswordSliсe';
-import './FormPasswordResetPage.scss';
-import { getIsUserAuth } from '../../services/selectors/userSelector';
+import './FormPasswordReset.scss';
 
-export const FormPasswordResetPage = () => {
+export default function FormPasswordReset() {
 	const dispatch = useDispatch();
-	const navigate = useNavigate();
-	const isConfirmPassword = useSelector(
-		(state) => state.resetPassword.isConfirmPassword
-	);
 	const { uid, token } = useParams();
-	const isUserAuth = useSelector(getIsUserAuth);
 
 	const validate = (values) => {
 		const errors = {};
@@ -25,11 +18,6 @@ export const FormPasswordResetPage = () => {
 		}
 		return errors;
 	};
-
-	const navigateHome = () => {
-		navigate('/');
-	};
-
 	const handleSubmit = (values) => {
 		dispatch(
 			fetchConfirmPasswordReset({
@@ -41,46 +29,18 @@ export const FormPasswordResetPage = () => {
 	};
 
 	return (
-		<main className="reset-password-form__container">
-			{isConfirmPassword ? (
-				<div className="reset-password-form__success-container">
-					<h2 className="reset-password-form__title">Пароль изменен</h2>
-					<p>Установлен новый пароль для учетной записи</p>
-					{isUserAuth ? (
-						<button
-							className="reset-password-form__button"
-							type="button"
-							onClick={navigateHome}
-						>
-							На главную
-						</button>
-					) : (
-						<button
-							className="reset-password-form__button"
-							type="button"
-							onClick={() => dispatch(openModal('login'))}
-						>
-							Войти
-						</button>
-					)}
-				</div>
-			) : (
-				<Formik
-					initialValues={{
-						newPassword: '',
-						confirmPassword: '',
-					}}
-					onSubmit={handleSubmit}
-					validate={validate}
-				>
-					{() => <FormComponent />}
-				</Formik>
-			)}
-		</main>
+		<Formik
+			initialValues={{
+				newPassword: '',
+				confirmPassword: '',
+			}}
+			onSubmit={handleSubmit}
+			validate={validate}
+		>
+			{() => <FormComponent />}
+		</Formik>
 	);
-};
-
-export default FormPasswordResetPage;
+}
 
 function FormComponent() {
 	const isLoadingConfirmPassword = useSelector(
