+++ src/pages/Area/AreaApp.scss
