+++ src/pages/Profile/PasswordData/PasswordData.scss
@@ -68,7 +68,7 @@ $input-border-radius: 8px;
 }
 
 .password-show-hide-btn-profile {
-	background: url(../../images/show-hide-password-button-profile.svg) no-repeat
+	background: url(../../../images/show-hide-password-button-profile.svg) no-repeat
 		center;
 	width: 16px;
 	height: 16px;
