+++ src/pages/Profile/PersonalData/PersonalData.scss
@@ -30,7 +30,7 @@ $input-border-radius: 8px;
 }
 
 .profile__avatar-button {
-	background: url(../../images/user-icon-avatar.jpg) center no-repeat;
+	background: url(../../../images/user-icon-avatar.jpg) center no-repeat;
 	margin: 0;
 	padding: 0;
 	border: 0;
